% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add_Matern}
\alias{add_Matern}
\title{Add Matern to formula}
\usage{
add_Matern(formula)
}
\arguments{
\item{formula}{a formula}
}
\value{
a formula
}
\description{
Add Matern to formula
}
\examples{
f <- staff_rangers_log ~ pop_density_log + PA_area_log + long
add_Matern(f)

}
