% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{aggregate_metrics}
\alias{aggregate_metrics}
\title{Aggregate the accuracy metrics}
\usage{
aggregate_metrics(metrics_df, fn = mean)
}
\arguments{
\item{metrics_df}{a data frame with metrics produced by \code{\link[=compute_metrics]{compute_metrics()}}}

\item{fn}{the function for aggregating all metrics but p-values}
}
\value{
a data frame
}
\description{
Aggregate the accuracy metrics
}
