% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{build_initial_trainning_data}
\alias{build_initial_trainning_data}
\title{Build the initial training datasets}
\usage{
build_initial_trainning_data(data, response, survey)
}
\arguments{
\item{data}{the dataset with the ranger data}

\item{response}{the unquoted name of the response variable}

\item{survey}{the criterion used to select rows depending on whether the focal number of staff is:
- completely unknown ("complete_unknown")
- completely or partially unknown ("partial_unknown")
- completely or partially known ("partial_known")
- completely known ("complete")
according to the choice, the variable PA_area is also adjusted.}
}
\value{
a tibble
}
\description{
This function creates a subset of the dataset \code{\link{data_rangers}} and log-transform (+1) some of its columns.
It contains no missing data for predictors.
}
\examples{
\dontrun{
## Here is how we created the data stored in this package:
data_test <- build_initial_trainning_data(data_rangers,
                                          response = staff_rangers,
                                          survey = "partial_known")
data_test <- data_test[!is.na(data_test$staff_rangers_log), ]
if (require(usethis)) {
  usethis::use_data(data_test, overwrite = TRUE)
}
}

}
