% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{build_training_data}
\alias{build_training_data}
\alias{build_initial_training_data,}
\alias{build_final_training_data}
\alias{build_initial_training_data}
\title{Build the training datasets}
\usage{
build_initial_training_data(data, response, survey)

build_final_training_data(data, formula, survey)
}
\arguments{
\item{data}{the complete dataset}

\item{response}{the unquoted name of the response variable}

\item{survey}{the criterion used to select rows depending on whether the focal number of staff is:
- completely unknown ("complete_unknown")
- completely or partially unknown ("partial_unknown")
- completely or partially known ("partial_known")
- completely known ("complete")
according to the choice, the variable PA_area is also adjusted.}

\item{formula}{the formula for the LMM or RF}
}
\value{
a tibble
}
\description{
These functions create a subset of the dataset \code{\link{data_rangers}} and log-transform (+1) some of its columns.
}
\section{Functions}{
\itemize{
\item \code{build_initial_training_data}: build the initial training datasets

\item \code{build_final_training_data}: build the final training datasets
}}

\examples{
\dontrun{
## Here is how we created the data stored in this package:
data_test <- build_initial_training_data(data_rangers,
                                          response = staff_rangers,
                                          survey = "partial_known")
data_test <- data_test[!is.na(data_test$staff_rangers_log), ]
if (require(usethis)) {
  usethis::use_data(data_test, overwrite = TRUE)
}
}

}
