% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_losses}
\alias{check_losses}
\title{Tool function to compute missing values and missing PA}
\usage{
check_losses(
  vars,
  data,
  resp_var = "staff_rangers_log",
  PA_var = "PA_area_log",
  PA_invlink = exp
)
}
\arguments{
\item{vars}{a vector of quoted variable names}

\item{data}{a data.frame or tibble}

\item{resp_var}{the name of the response variable (to exclude it from computation)}

\item{PA_var}{the name of the variable storing the PA area information}

\item{PA_invlink}{the function to turn PA_var into km2 (default = exp)}
}
\value{
a tibble
}
\description{
Tool function to compute missing values and missing PA
}
