% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clipout_polygon}
\alias{clipout_polygon}
\title{Remove polygon at a given location from a set of polygons}
\usage{
clipout_polygon(data, countryname_eng, lon, lat)
}
\arguments{
\item{data}{A dataset containing at least the columns \code{geometry} and \code{name} or \code{rne_name}}

\item{countryname_eng}{The name of a country in English}

\item{lon}{A vector of 2 longitude coordinates defining a box containing the polygons to extract}

\item{lat}{A vector of 2 latitude coordinates defining a box containing the polygons to extract}
}
\description{
This functions is a helper function to be used in combination to \code{\link[=extract_polygon]{extract_polygon()}}.
It removes a set of polygons nested within a multipolygon sf geometry.
}
\examples{
## Clipping out the polygon for French Guiana
world <- rnaturalearth::ne_countries(scale = "medium", returnclass = "sf")
France <- world$geometry[world$name == "France"]
plot(France)
France2 <- clipout_polygon(data = world,
                           countryname_eng = "France",
                           lon = c(-55, -50), lat = c(6, 2))
plot(France2)

}
