% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compute_distance}
\alias{compute_distance}
\title{Compute the distance (or inverse distance) between locations}
\usage{
compute_distance(long, lat, inv = FALSE)
}
\arguments{
\item{long}{a vector of longitudes}

\item{lat}{a vector of latitudes}

\item{inv}{whether to return the inverse of the distance (default = FALSE)}
}
\value{
a square data frame (if inv = FALSE) or matrix (if inv = TRUE)
}
\description{
Compute the distance (or inverse distance) between locations
}
\examples{
# distance Berlin/London:
compute_distance(long = c(13.41, -0.13), lat = c(52.52, 51.51))

}
