% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compare predictions and observations using various metrics}
\usage{
compute_metrics(pred, obs, inv.dist = NULL)
}
\arguments{
\item{pred}{a vector of predicted values}

\item{obs}{a vector of observed values}

\item{inv.dist}{a matrix of inverse distance between points (optional but needed to compute Moran's I)}
}
\value{
a data frame with all the metrics
}
\description{
Compare predictions and observations using various metrics
}
\examples{
compute_metrics(pred = c(1.1, 1.8, 2.5, 4.5),
                obs = c(1, 2, 3, 4))

proxy <- compute_distance(long = c(13.41, -0.13, -77.04, 116.24),
                          lat = c(52.52, 51.51, 38.89, 39.90),
                          inv = TRUE)

compute_metrics(pred = c(1.1, 1.8, 2.5, 4.5),
                obs = c(1, 2, 3, 4),
                inv.dist = proxy)

}
