% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_rangers}
\alias{data_rangers}
\title{Ranger data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 249 rows and 37 columns.
}
\usage{
data_rangers
}
\description{
This object contain the data about rangers and non-rangers working in protected areas.
}
\examples{
data_rangers

}
\seealso{
\code{\link[=fetch_data]{fetch_data()}} for the function used to create such a dataset
}
\keyword{datasets}
