% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_rangers_with_geo}
\alias{data_rangers_with_geo}
\title{Ranger data with polygons}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 263 rows and 38 columns.
}
\usage{
data_rangers_with_geo
}
\description{
This object contain the data about rangers and non-rangers working in protected areas.
Compared to \code{\link{data_rangers}}, this dataset contains polygons corresponding to each country / territory.
It therefore contains additional rows for which we did not collect any information.
}
\examples{
data_rangers_with_geo

}
\seealso{
\code{\link[=fetch_data]{fetch_data()}} for the function used to create such a dataset
}
\keyword{datasets}
