% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_test}
\alias{data_test}
\title{Test data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120 rows and 17 columns.
}
\usage{
data_test
}
\description{
This object contain a subset with the data about rangers and other staffs members working in protected areas.
}
\examples{
data_test

}
\seealso{
\code{\link[=build_data_test]{build_data_test()}} for the function used to create such a dataset
}
\keyword{datasets}
