% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{drop_logs}
\alias{drop_logs}
\title{Drop logs from formula}
\usage{
drop_logs(formula)
}
\arguments{
\item{formula}{a formula}
}
\value{
a formula
}
\description{
Drop logs from formula
}
\examples{
f <- staff_rangers_log ~ pop_density_log + PA_area_log + long + Matern(1|long + lat)
drop_logs(f)

}
