% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_PA_areas}
\alias{extract_PA_areas}
\title{Internal function used to extract PA_areas at the level of countries/territories}
\usage{
extract_PA_areas(data_final_pred, resp)
}
\arguments{
\item{data_final_pred}{A final prediction dataset}

\item{resp}{The quoted name of the response variable}
}
\description{
This function is called by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}
}
