% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_polygon}
\alias{extract_polygon}
\title{Extract polygon at a given location from a set of polygons}
\usage{
extract_polygon(data, countryname_eng, lon, lat)
}
\arguments{
\item{data}{A dataset containing at least the columns \code{geometry} and \code{name} or \code{rne_name}}

\item{countryname_eng}{The name of a country in English}

\item{lon}{A vector of 2 longitude coordinates defining a box containing the polygons to extract}

\item{lat}{A vector of 2 latitude coordinates defining a box containing the polygons to extract}
}
\description{
This functions is a helper function used during the data preparation to extract polygons nested within a
multipolygon sf geometry. We use this because our data set makes distinction between territories that
are not down in the dataset we use to retrieve the polygons.
}
\examples{
## Extracting the polygon for French Guiana
world <- rnaturalearth::ne_countries(scale = "medium", returnclass = "sf")
FrenchGuiana <- extract_polygon(data = world,
                                countryname_eng = "France",
                                lon = c(-55, -50), lat = c(6, 2))
plot(FrenchGuiana)

}
