% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{extract_results}
\alias{extract_results}
\alias{extract_results_internal}
\title{Extract and format the results produced by the worflow}
\usage{
extract_results(list_results_LMM = list(), list_results_RF = list())

extract_results_internal(what, who, type)
}
\arguments{
\item{list_results_LMM}{a list of objects produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{list_results_RF}{a list of objects produced by \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{what}{an object produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{who}{the type of staff ("rangers", "others" or "all")}

\item{type}{the type of workflow ("LMM" or "RF")}
}
\description{
Extract and format the results produced by the worflow
}
\section{Functions}{
\itemize{
\item \code{extract_results_internal}: an internal funciton fetching the results
}}

\examples{
\dontrun{
res <- extract_results(
       list_results_LMM = list(LMM_000, LMM_025, LMM_050, LMM_075, LMM_100),
       list_results_RF  = list(RF_000, RF_025, RF_050, RF_075, RF_100))
}

}
