% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{extract_training_info}
\alias{extract_training_info}
\alias{extract_training_info_internal}
\title{Extract training data info}
\usage{
extract_training_info(
  which,
  list_results_LMM = list(),
  list_results_RF = list(),
  data = NULL
)

extract_training_info_internal(which, what, who, type, data)
}
\arguments{
\item{which}{either \code{"initial"} or \code{"final"} depending on which training dataset to summarise}

\item{list_results_LMM}{a list of objects produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{list_results_RF}{a list of objects produced by \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}

\item{what}{an object produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{who}{the type of staff ("rangers", "others" or "all")}

\item{type}{the type of workflow ("LMM" or "RF")}
}
\description{
Extract training data info
}
\section{Functions}{
\itemize{
\item \code{extract_training_info_internal()}: an internal function fetching the info on training datasets

}}
\examples{
\dontrun{
LMM_small_test <- run_LMM_workflow(data = data_rangers, Ncpu = 2, coef = 0,
                                     rep_feature_select = 2, rep_finetune = 2, rep_simu = 2)

RF_small_test <- run_RF_workflow(data = data_rangers, Ncpu = 2, coef = 0,
                                 rep_feature_select = 2, rep_finetune = 2, rep_simu = 2,
                                 grid_type = "coarse",
                                 n_trees = 100)

extract_training_info(list_results_LMM = list(LMM_small_test),
                list_results_RF  = list(RF_small_test))
}

}
