% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection_LMM}
\alias{feature_selection_LMM}
\title{Perform feature selection on LMM}
\usage{
feature_selection_LMM(
  full_fit,
  data,
  metric = "RMSE",
  minimise = TRUE,
  rep = 10,
  Ncpu = 1,
  target = "staff_rangers_log",
  spatial = "Matern",
  seed = 123,
  ...
)
}
\arguments{
\item{full_fit}{a full fitted model}

\item{data}{the full dataset}

\item{metric}{the metric used for computing prediction accuracy (see \code{\link[=compute_metrics]{compute_metrics()}})}

\item{minimise}{whether the metric should be minimise (TRUE, default) or maximise (FALSE)}

\item{rep}{the number of cross validation replicates (default = 10)}

\item{Ncpu}{the number of CPU cores to be used (default = 1)}

\item{target}{the quoted name of the response variable (default = "staff_rangers_log")}

\item{spatial}{either FALSE (default) or "Matern"}

\item{seed}{the seed used to control the reproducibility of the cross validation}

\item{...}{additional parameters to be passed to \code{\link[spaMM:fitme]{spaMM::fitme()}}}
}
\value{
a list with metrics and best formula
}
\description{
Perform feature selection on LMM
}
