% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch the data from the Google sheet}
\usage{
fetch_data(keep_geometry = FALSE)
}
\arguments{
\item{keep_geometry}{whether or not to keep a geometry column storing the polygons associated with each country/territory (default = \code{FALSE})}
}
\value{
a tibble with the raw data
}
\description{
This function retrieves the raw data from a Google spreadsheet and format them.
}
\examples{
\dontrun{
## Here is how we created the data stored in this package:

data_rangers <- fetch_data(keep_geometry = FALSE)
data_rangers_with_geo <- fetch_data(keep_geometry = TRUE)

write.csv2(data_rangers, file = "inst/extdata/raw_data/data_rangers.csv", row.names = FALSE)

if (require(usethis)) {
  usethis::use_data(data_rangers, overwrite = TRUE)
  usethis::use_data(data_rangers_with_geo, overwrite = TRUE)
}
}

}
