% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fill_PA_area}
\alias{fill_PA_area}
\title{Impute missing personnel numbers proportionally to the known density in surveyed areas}
\usage{
fill_PA_area(data, coef)
}
\arguments{
\item{data}{a data.frame or tibble produced by \code{\link[=fetch_data]{fetch_data()}}}

\item{coef}{the coefficient used for the propagation}
}
\value{
a dataframe or tibble
}
\description{
This function modifies the numbers of personnel for countries/territories where it is partially known.
For those countries/territories, it assumes that the unsurveyed area have a density of personnel proportional to the density in the surveyed areas.
The proportionality is set by the argument \code{coef}:
- e.g. if \code{coef = 1}, then unsurveyed areas are populated with the same density as the surveyed areas.
- e.g. if \code{coef = 0.5}, then unsurveyed areas are populated with half the density as the surveyed areas.
}
\examples{
data_rangers_50 <- fill_PA_area(data_rangers, coef = 0.5)
surveyed_fraction <- with(data_rangers, PA_area_surveyed /
     (PA_area_surveyed + PA_area_unsurveyed))
plot(surveyed_fraction,
     surveyed_fraction * data_rangers_50$staff_rangers / data_rangers$staff_rangers,
     ylim = c(0, 1), xlim = c(0, 1))
polygon(x = c(0, 1, 0, 0), y = c(0, 1, 0.5, 0))

}
