% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tunning_LMM.R
\name{finetune_LMM}
\alias{finetune_LMM}
\title{Fine tune LMM}
\usage{
finetune_LMM(formula, data, rep = 10, Ncpu = 1, seed = 123)
}
\arguments{
\item{formula}{a model formula used to fit the data}

\item{data}{the full dataset}

\item{rep}{the number of cross validation replicates (default = 10)}

\item{Ncpu}{the number of CPU cores to be used (default = 1)}

\item{seed}{the seed used to control the reproducibility of the cross validation}
}
\description{
Fine tune LMM
}
