% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{formula_top_pred_LMM}
\alias{formula_top_pred_LMM}
\title{Build formula from top predictors}
\usage{
formula_top_pred_LMM(fit, k = NULL)
}
\arguments{
\item{fit}{a model fitted with \href{spaMM::fitme}{\code{fitme()}}}

\item{k}{the maximum number of predictors to return (default = NULL for no sub-selection)}
}
\value{
a formula
}
\description{
Build formula from top predictors
}
