% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{package_info}
\alias{package_info}
\title{Provide some basic information about the code of the package}
\usage{
package_info()
}
\description{
This is a function for development purpose only. This function provides the
number of functions and lines of code of the package. The results depend on
whether the function is called from the compiled or the raw package. When
called from the compiled package, the number of functions only gives the
number of exported functions, and the number of lines of code no longer
includes documentation and examples.
}
\note{
This function does not work with \code{devtools::load_all(".")}.
}
