% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_density_staff}
\alias{plot_density_staff}
\title{Plot densities of rangers or all personnel}
\usage{
plot_density_staff(
  what,
  who,
  data,
  ymax = 6000,
  breaks = c(10^(0:3), 2 * 10^(0:3), 5 * 10^(0:3)),
  tag = "",
  basesize = 7
)
}
\arguments{
\item{what}{an object produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{who}{the type of staff ("rangers", "others" or "all")}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}

\item{ymax}{the maximal value for the y-axis}

\item{breaks}{a vector of number defining the horizontal lines in the plot (i.e. breaks in ggplot jargon)}

\item{tag}{a tag to be used for building a panel (e.g "A.")}

\item{basesize}{the value for the argument \code{base_size} used in \href{ggplot2::theme}{\code{ggplot2::theme()}}}
}
\description{
Plot densities of rangers or all personnel
}
\examples{
# see ?rangeRinPA

}
