% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_density_vs_sampling}
\alias{plot_density_vs_sampling}
\title{Plot the density of staff against the sampling intensity}
\usage{
plot_density_vs_sampling(data, who = "rangers", coef = 1)
}
\arguments{
\item{data}{the complete dataset}

\item{who}{\code{"rangers"} (default), \code{"others"} or \code{"all"}}

\item{coef}{the coefficient for the imputation}
}
\description{
Plot the density of staff against the sampling intensity
}
\examples{
plot_density_vs_sampling(data_rangers)

}
