% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_features_selected}
\alias{plot_features_selected}
\title{Plot the output of the selection of predictors fits}
\usage{
plot_features_selected(list_results_LMM, list_results_RF, data, size = 4)
}
\arguments{
\item{list_results_LMM}{a list of objects produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{list_results_RF}{a list of objects produced by \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}

\item{size}{the size of the points on the plot}
}
\description{
Plot the output of the selection of predictors fits
}
\examples{
# see ?extract_results for example

}
