% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_features_selection}
\alias{plot_features_selection}
\title{Plot the influence of the selection of predictors on the RMSE}
\usage{
plot_features_selection(result, who = "rangers")
}
\arguments{
\item{result}{the output of a call to \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{who}{\code{"rangers"} (default), \code{"others"} or \code{"all"}}
}
\description{
Plot the influence of the selection of predictors on the RMSE
}
\examples{
\dontrun{
## Note: this example takes some time to run!
RF_test <- run_RF_workflow(data = data_rangers, Ncpu = 2, coef = 0,
                           rep_feature_select = 2, rep_finetune = 10, rep_simu = 2,
                           grid_type = "fine", n_trees = 100)

plot_features_selection(result = RF_test, who = "rangers")
}

}
