% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_features_selection_panel}
\alias{plot_features_selection_panel}
\title{Plot the influence of the selection of predictors on the RMSE}
\usage{
plot_features_selection_panel(result1, result2, who = "rangers")
}
\arguments{
\item{result1}{the first output of a call to \code{\link[=run_RF_workflow]{run_RF_workflow()}} or \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{result2}{the second output of a call to \code{\link[=run_RF_workflow]{run_RF_workflow()}} or \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{who}{\code{"rangers"} (default), \code{"others"} or \code{"all"}}
}
\description{
Plot the influence of the selection of predictors on the RMSE
}
\examples{
# see see ?rangeRinPA

}
\seealso{
\link{plot_features_selection}
}
