% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_finetuning}
\alias{plot_finetuning}
\title{Plot the output of the fine tuning for RF/ETs fits}
\usage{
plot_finetuning(result, who = "rangers")
}
\arguments{
\item{result}{the output of a call to \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{who}{\code{"rangers"} (default), \code{"others"} or \code{"all"}}
}
\description{
Plot the output of the fine tuning for RF/ETs fits
}
\examples{
\dontrun{
## Note: this example takes some time to run!
RF_test <- run_RF_workflow(data = data_rangers, Ncpu = 2, coef = 0,
                           rep_feature_select = 2, rep_finetune = 10, rep_simu = 2,
                           grid_type = "fine", n_trees = 100)

plot_finetuning(result = RF_test, who = "rangers")
}

}
