% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_map_reliability}
\alias{plot_map_reliability}
\title{Plot the map for the reliability score}
\usage{
plot_map_reliability(data, proj = "+proj=moll")
}
\arguments{
\item{data}{the complete dataset}

\item{proj}{the projection to be used (e.g. "+proj=moll" - the default)}
}
\description{
Plot the map for the reliability score
}
\examples{
plot_map_reliability(data_rangers_with_geo)

}
