% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_PA_by_data_type}
\alias{plot_PA_by_data_type}
\title{Plot details about estimations in terms of PA}
\usage{
plot_PA_by_data_type(what, data)
}
\arguments{
\item{what}{an object produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}
}
\description{
Plot details about estimations in terms of PA
}
\examples{
# see ?rangeRinPA

}
