% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_projections}
\alias{plot_projections}
\title{Plot projections for required numbers}
\usage{
plot_projections(what, data)
}
\arguments{
\item{what}{an object produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{the complete dataset}
}
\description{
Plot projections for required numbers
}
\examples{
# see ?rangeRinPA

}
