% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_reliability_vs_sampling}
\alias{plot_reliability_vs_sampling}
\title{Plot the reliability score against the sampling intensity}
\usage{
plot_reliability_vs_sampling(data)
}
\arguments{
\item{data}{the complete dataset}
}
\description{
Plot the reliability score against the sampling intensity
}
\examples{
plot_reliability_vs_sampling(data_rangers)

}
