% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prepare_grid_finetuning}
\alias{prepare_grid_finetuning}
\title{Internal function used to prepare the grid of the parameters to try when finetuning the RF}
\usage{
prepare_grid_finetuning(grid_type = "fine")
}
\arguments{
\item{grid_type}{The type of grid to use for finetuning the RF ("fine", the default, or "coarse")}
}
\description{
This function is called by \code{\link[=run_RF_workflow]{run_RF_workflow()}}.
}
\examples{
nrow(prepare_grid_finetuning())
nrow(prepare_grid_finetuning(grid_type = "coarse"))

}
