% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_LMM.R
\name{rank_predictors_LMM}
\alias{rank_predictors_LMM}
\title{Extract top predictors}
\usage{
rank_predictors_LMM(fit, k = NULL)
}
\arguments{
\item{fit}{a model fitted with \href{spaMM::fitme}{\code{fitme()}}}

\item{k}{the maximum number of predictors to return (default = NULL for no sub-selection)}
}
\value{
predictors with their absolute t-values
}
\description{
This function extract the top (fixed-effects) predictors from a LMM based on their absolute t-value.
}
