% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{RMSE}
\alias{RMSE}
\title{Compute root mean square error}
\usage{
RMSE(pred, obs)
}
\arguments{
\item{pred}{a vector of predictions}

\item{obs}{a vector of observations}
}
\description{
Compute root mean square error
}
