% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_RF.R
\name{run_RF_workflow}
\alias{run_RF_workflow}
\title{Run the full workflow for RF}
\usage{
run_RF_workflow(
  data,
  rerank = TRUE,
  Ncpu = 2,
  coef = 1,
  rep_feature_select = 1000,
  rep_finetune = 1000,
  rep_simu = 10000,
  n_trees = 10000,
  grid_type = "fine",
  outliers = "Greenland"
)
}
\arguments{
\item{data}{the full dataset}

\item{rerank}{whether or not to recompute variable importance recursively during selection (default = TRUE)}

\item{Ncpu}{the number of CPU cores to be used (default = 1)}

\item{coef}{the coefficient used to population rangers in unsurveyed part of surveyed countries (see \code{\link[=fill_PA_area]{fill_PA_area()}})}

\item{rep_feature_select}{the number of replicates for the feature selection (default = 1000)}

\item{rep_finetune}{the number of replicates for fine tuning (default = 1000)}

\item{rep_simu}{the number of simulation replicates (default = 10000)}

\item{n_trees}{the number of trees in the random forest}

\item{grid_type}{The type of grid to use for finetuning the RF ("fine", the default, or "coarse")}

\item{outliers}{a vector with the names of the countries/territories to discard (default = \verb{"Greenland})}
}
\value{
a list with all the output information
}
\description{
Run the full workflow for RF
}
\examples{
\dontrun{
  RF_small_test <- run_RF_workflow(data = data_rangers, Ncpu = 2, coef = 1,
                                   rep_feature_select = 2, rep_finetune = 2, rep_simu = 2,
                                   grid_type = "coarse", n_trees = 100)
}

}
