% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{single_summary_internal}
\alias{single_summary_internal}
\title{An internal function fetching the results at the country level}
\usage{
single_summary_internal(result, who, resp, data)
}
\arguments{
\item{result}{a result table produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{who}{the type of staff ("rangers", "others" or "all")}

\item{resp}{the quoted name of the response variable}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}
}
\description{
An internal function fetching the results at the country level
}
\examples{
\dontrun{
  LMM_small_test <- run_LMM_workflow(data = data_rangers, Ncpu = 2, coef = 0,
                                     rep_feature_select = 2, rep_finetune = 2, rep_simu = 2)
  single_summary_internal(result = LMM_small_test,
                          who = "rangers", resp = "staff_rangers_log", data = data_rangers)
}

}
