% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_completeness_km2}
\alias{table_completeness_km2}
\title{Create completeness table (km2)}
\usage{
table_completeness_km2(data, outliers = "GRL")
}
\arguments{
\item{data}{the complete dataset}

\item{outliers}{a vector with the iso code for the countries/territories to discard (default = \code{"GRL"})}
}
\value{
a tibble
}
\description{
This function creates the table that indicate the surface of protected areas with ranger staff and other staff documented in our data.
}
\examples{
table_completeness_km2(data_rangers)

}
