% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_completeness_obs}
\alias{table_completeness_obs}
\title{Create completeness table (observations)}
\usage{
table_completeness_obs(data)
}
\arguments{
\item{data}{the complete dataset created with \href{fetch_data}{\code{fetch_data}()}.}
}
\value{
a tibble
}
\description{
This function creates the table that indicate how many countries and territories have ranger staff and other staff documented in our data.
}
\examples{
table_completeness_obs(data_rangers)

}
