% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_predictions_per_continent}
\alias{table_predictions_per_continent}
\title{Create table with predictions (per continent)}
\usage{
table_predictions_per_continent(what, data)
}
\arguments{
\item{what}{an object produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}
}
\value{
a tibble
}
\description{
This function creates the table that shows the point predictions and their intervals, for the
different continents.
NOTE: superseeded by \href{table_predictions_summary}{\code{table_predictions_summary}()}.
}
\examples{
# see see ?rangeRinPA

}
