% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_predictions_per_method}
\alias{table_predictions_per_method}
\title{Create table with predictions (per method)}
\usage{
table_predictions_per_method(
  list_results_LMM,
  list_results_RF,
  data,
  density = FALSE
)
}
\arguments{
\item{list_results_LMM}{a list of objects produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{list_results_RF}{a list of objects produced by \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}

\item{density}{a logical indicating whether to return densities instead of raw numbers (default =
\code{FALSE})}
}
\value{
a tibble
}
\description{
This function creates the table that shows the point predictions and their intervals, for the
different methods and coefficient values used for the imputation.
}
\examples{
# see see ?rangeRinPA

}
