% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_projections}
\alias{table_projections}
\title{This function creates the table that shows the requirements in rangers and all personal.
IIt creates of the main text tables.}
\usage{
table_projections(what, data)
}
\arguments{
\item{what}{an object produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}
}
\value{
a tibble
}
\description{
This function creates the table that shows the requirements in rangers and all personal.
IIt creates of the main text tables.
}
\examples{
# see see ?rangeRinPA

}
