% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_raw_data}
\alias{table_raw_data}
\title{Create raw data table}
\usage{
table_raw_data(data)
}
\arguments{
\item{data}{the complete dataset created with \link[=`fetch_data`]{\code{fetch_data}()}.}
}
\value{
a tibble
}
\description{
This function creates the table with the core information about surveyed countries/territories.
}
\examples{
table_raw_data(data_rangers)

}
