% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_training_final}
\alias{table_training_final}
\title{Create table with info on final training datasets}
\usage{
table_training_final(list_results_LMM, list_results_RF, data)
}
\arguments{
\item{list_results_LMM}{a list of objects produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{list_results_RF}{a list of objects produced by \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}
}
\value{
a tibble
}
\description{
This function creates the table that shows the information about the final training datasets.
}
\examples{
# see see ?rangeRinPA

}
