% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{totitle}
\alias{totitle}
\title{Mixed case capitalizing}
\usage{
totitle(x)
}
\arguments{
\item{x}{a string of characters}
}
\value{
a string of characters
}
\description{
This function turn the first letter of each word into a capital letter.
}
\examples{
totitle("once upon a time")

}
