% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_LMM}
\alias{validate_LMM}
\title{Compute accuracy metrics on test set for LMM}
\usage{
validate_LMM(
  formula,
  data,
  rep = 10,
  Ncpu = 1,
  target = "staff_rangers_log",
  spatial = FALSE,
  seed = 123,
  return_fit = TRUE,
  ...
)
}
\arguments{
\item{formula}{a model formula used to fit the data}

\item{data}{the full dataset}

\item{rep}{the number of cross validation replicates (default = 10)}

\item{Ncpu}{the number of CPU cores to be used (default = 1)}

\item{target}{the quoted name of the response variable (default = "staff_rangers_log")}

\item{spatial}{either FALSE (default) or TRUE (or "Matern" as synonymous)}

\item{seed}{the seed used to control the reproducibility of the cross validation}

\item{return_fit}{whether to return the fit on all the data as attribute (default = FALSE)}

\item{...}{additional parameters to be passed to \code{\link[spaMM:fitme]{spaMM::fitme()}}}
}
\value{
a tibble with the CV replicates in row and accuracy metrics in columns
}
\description{
Compute accuracy metrics on test set for LMM
}
\examples{
validate_LMM(staff_rangers_log ~ PA_area_log, data = data_test)

}
\seealso{
\code{\link[=compute_metrics]{compute_metrics()}} for details on the function computing the values outputted here
}
