% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_RF}
\alias{validate_RF}
\title{Compute accuracy metrics on test set for RF}
\usage{
validate_RF(
  formula,
  data,
  rep = 10,
  Ncpu = 1,
  spatial = FALSE,
  seed = 123,
  method = "CV",
  return_fit = TRUE,
  ...
)
}
\arguments{
\item{formula}{a model formula used to fit the data}

\item{data}{the full dataset}

\item{rep}{the number of cross validation replicates (default = 10)}

\item{Ncpu}{the number of CPU cores to be used (default = 1)}

\item{spatial}{either FALSE (default), or TRUE (to use matrix of distances as predictor)}

\item{seed}{the seed used to control the reproducibility of the cross validation}

\item{method}{either "CV" for cross-validation or "OOB" for directly using the out-of-bag observations generated when growing the forest}

\item{return_fit}{whether to return the fit on all the data as attribute (default = FALSE)}

\item{...}{additional parameters to be passed to \code{\link[ranger:ranger]{ranger::ranger()}}}
}
\value{
a tibble with the CV/OOB replicates in row and accuracy metrics in columns
}
\description{
Compute accuracy metrics on test set for RF
}
\examples{
validate_RF(staff_rangers_log ~ PA_area_log, data = data_test, method = "CV")
validate_RF(staff_rangers_log ~ PA_area_log, data = data_test,
                 rep = 1, num.trees = 10, method = "OOB")


}
\seealso{
\code{\link[=compute_metrics]{compute_metrics()}} for details on the function computing the values outputted here
}
