% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{build_data_test}
\alias{build_data_test}
\title{Build the test dataset}
\usage{
build_data_test(data)
}
\arguments{
\item{data}{the dataset with the ranger data}
}
\value{
a tibble
}
\description{
This function creates a subset of the dataset \code{\link{data_rangers}} and log-transform (+1) some of its columns.
It contains no missing data.
}
\examples{
\dontrun{
## Here is how we created the data stored in this package:
data_test <- build_data_test(data_rangers)
if (require(usethis)) {
  usethis::use_data(data_test, overwrite = TRUE)
}
}

}
