% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_PA_areas}
\alias{extract_PA_areas}
\title{Internal function used to extract PA_areas at the level of countries/territories}
\usage{
extract_PA_areas(data_final_pred, resp, data)
}
\arguments{
\item{data_final_pred}{A final prediction dataset}

\item{resp}{The quoted name of the response variable}

\item{data}{the full dataset}
}
\description{
This function is called by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}.
}
\examples{
dat <- fill_PA_area(data_rangers, coef = 0.5)
d <- build_final_pred_data(data = dat,
                           formula = staff_rangers_log ~ GDP_2019,
                           survey = "complete_known",
                           spatial = TRUE)
d$data_predictable$staff_rangers_log_predicted <- 0
extract_PA_areas(d, "staff_rangers_log", dat)

}
