% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{extract_results}
\alias{extract_results}
\alias{extract_results_internal}
\title{Extract and format the results produced by the workflow}
\usage{
extract_results(
  list_results_LMM = list(),
  list_results_RF = list(),
  data = NULL
)

extract_results_internal(what, who, type, data)
}
\arguments{
\item{list_results_LMM}{a list of objects produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{list_results_RF}{a list of objects produced by \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}

\item{what}{an object produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}} or \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{who}{the type of staff ("rangers", "others" or "all")}

\item{type}{the type of workflow ("LMM" or "RF")}
}
\description{
Extract and format the results produced by the workflow
}
\section{Functions}{
\itemize{
\item \code{extract_results_internal()}: an internal function fetching the results

}}
\examples{
\dontrun{
LMM_small_test <- run_LMM_workflow(data = data_rangers, Ncpu = 2, coef = 1,
                                   rep_feature_select = 2, rep_finetune = 2, rep_simu = 2)

RF_small_test <- run_RF_workflow(data = data_rangers, Ncpu = 2, coef = 1,
                                 rep_feature_select = 2, rep_finetune = 2, rep_simu = 2,
                                 grid_type = "coarse",
                                 n_trees = 100)

extract_results(list_results_LMM = list(LMM_small_test),
                list_results_RF  = list(RF_small_test)) \%>\%
  tidyr::unnest_wider(PA_areas)

plot_features_selected(list_results_LMM = list(LMM_small_test),
                       list_results_RF  = list(RF_small_test), data = data_rangers)

}

}
