% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection_LMM}
\alias{feature_selection_LMM}
\alias{feature_selection_LMM,}
\alias{feature_selection_LMM_internal}
\title{Perform feature selection on LMMs}
\usage{
feature_selection_LMM(
  full_fit,
  metric = "RMSE",
  minimise = TRUE,
  rep = 10,
  Ncpu = 1,
  target = "staff_rangers_log",
  seed = 123,
  ...
)

feature_selection_LMM_internal(
  full_fit,
  rep = 10,
  Ncpu = 1,
  target = "staff_rangers_log",
  spatial = "Matern",
  seed = 123,
  ...
)
}
\arguments{
\item{full_fit}{a full fitted model}

\item{metric}{the metric used for computing prediction accuracy (see \code{\link[=compute_metrics]{compute_metrics()}})}

\item{minimise}{whether the metric should be minimise (TRUE, default) or maximise (FALSE)}

\item{rep}{the number of cross validation replicates (default = 10)}

\item{Ncpu}{the number of CPU cores to be used (default = 1)}

\item{target}{the quoted name of the response variable (default = "staff_rangers_log")}

\item{seed}{the seed used to control the reproducibility of the cross validation}

\item{...}{additional parameters to be passed to \code{\link[spaMM:fitme]{spaMM::fitme()}}}

\item{spatial}{either FALSE (default) or "Matern"}
}
\description{
Perform feature selection on LMMs
}
\section{Functions}{
\itemize{
\item \code{feature_selection_LMM}: wrapper function performing the feature selection on LMMs with and without the Matern term

\item \code{feature_selection_LMM_internal}: internal function performing the feature selection on LMMs
}}

