% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fetch_data_rangers}
\alias{fetch_data_rangers}
\title{Fetch the ranger data from the Google sheet}
\usage{
fetch_data_rangers()
}
\value{
a tibble with the raw data
}
\description{
This function retrieves the data from a Google spreadsheet and format them
}
\examples{
\dontrun{
## Here is how we created the data stored in this package:
data_rangers <- fetch_data_rangers()
if (require(usethis)) {
  usethis::use_data(data_rangers)
}
}

}
