% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tunning_RF.R
\name{finetune_RF}
\alias{finetune_RF}
\title{Fine tune the parameters of a random forest}
\usage{
finetune_RF(
  values_to_try,
  param_to_tune,
  formula,
  data,
  rep = 10,
  Ncpu = 1,
  target = "staff_rangers_log",
  fn = mean,
  ...
)
}
\arguments{
\item{values_to_try}{the parameter value to try}

\item{param_to_tune}{the quoted name of the parameter to study}

\item{formula}{the formula for the forest}

\item{data}{the full dataset}

\item{rep}{the number of cross validation replicates (default = 10)}

\item{Ncpu}{the number of CPU cores to be used (default = 1)}

\item{target}{the quoted name of the response variable (default = "staff_rangers_log")}

\item{fn}{the function for aggregating all metrics but p-values}

\item{...}{additional parameters to be passed to \code{\link[ranger:ranger]{ranger::ranger()}}}
}
\value{
a list with a data frame and a plot
}
\description{
Fine tune the parameters of a random forest
}
