% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_density_vs_PA}
\alias{plot_density_vs_PA}
\title{Plot densities of staff vs area PA}
\usage{
plot_density_vs_PA(data, who, coef = 1)
}
\arguments{
\item{data}{the complete dataset}

\item{who}{\code{"rangers"} (default), \code{"others"} or \code{"all"}}

\item{coef}{the coefficient for the imputation}
}
\description{
Plot densities of staff vs area PA
}
\examples{
plot_density_vs_PA(data = data_rangers, who = "rangers", coef = 1)

}
