% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_density_vs_PA_panel}
\alias{plot_density_vs_PA_panel}
\title{Plot densities of personnel and rangers vs area PA}
\usage{
plot_density_vs_PA_panel(data, coef, basesize = 7)
}
\arguments{
\item{data}{the complete dataset}

\item{coef}{the coefficient for the imputation}

\item{basesize}{the value for the argument \code{base_size} used in \href{ggplot2::theme}{\code{ggplot2::theme()}}}
}
\description{
Plot densities of personnel and rangers vs area PA
}
\examples{
plot_density_vs_PA_panel(data = data_rangers, coef = 1)

}
