% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_map_sampling}
\alias{plot_map_sampling}
\title{Plot the map of the sampling design}
\usage{
plot_map_sampling(data, proj = "+proj=moll")
}
\arguments{
\item{data}{the complete dataset}

\item{proj}{the projection to be used (e.g. "+proj=moll" - the default)}
}
\description{
Plot the map of the sampling design
}
\examples{
plot_map_sampling(data_rangers)

if(require(patchwork)) {
  plot_map_sampling(data_rangers, proj = "+proj=moll") +
  plot_map_sampling(data_rangers, proj = "+proj=robin" ) +
  plot_map_sampling(data_rangers, proj = "+proj=wintri") +
  plot_map_sampling(data_rangers, proj = "+proj=natearth2") +
  plot_map_sampling(data_rangers, proj = "+proj=mbt_fps") +
  plot_map_sampling(data_rangers, proj = "+proj=hammer") +
  plot_map_sampling(data_rangers, proj = "+proj=wag1") +
  plot_map_sampling(data_rangers, proj = "+proj=goode") +
  plot_map_sampling(data_rangers, proj = "+proj=eqearth") +
  plot_map_sampling(data_rangers, proj = "+proj=eck4") +
  plot_map_sampling(data_rangers, proj = "+proj=boggs") +
  plot_map_sampling(data_rangers, proj = "+proj=aitoff")
}

}
