% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_tallies_across_methods}
\alias{plot_tallies_across_methods}
\title{Plot tallies across methods}
\usage{
plot_tallies_across_methods(
  list_results_LMM,
  list_results_RF,
  data,
  basesize = 7
)
}
\arguments{
\item{list_results_LMM}{a list of objects produced by \code{\link[=run_LMM_workflow]{run_LMM_workflow()}}}

\item{list_results_RF}{a list of objects produced by \code{\link[=run_RF_workflow]{run_RF_workflow()}}}

\item{data}{a dataset with info on continents if breakdown by continent is required (optional)}

\item{basesize}{the value for the argument \code{base_size} used in \href{ggplot2::theme}{\code{ggplot2::theme()}}}
}
\description{
Plot tallies across methods
}
\examples{
# see ?rangeRinPA

}
