% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prepare_data}
\alias{prepare_data}
\title{Tool function to compute train and test datasets given a formula}
\usage{
prepare_data(formula, data, test_prop = 0)
}
\arguments{
\item{formula}{the formula for the LMM or RF}

\item{data}{the complete dataset}

\item{test_prop}{the amount of rows to keep in the test dataset (default = 0)}
}
\description{
Tool function to compute train and test datasets given a formula
}
\examples{
prepare_data(log(staff_rangers + 1) ~ log(GDP_2019), data = data_rangers, test_prop = 0.1)

}
