% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prepare_vars}
\alias{prepare_vars}
\title{Identify variables available for focal countries}
\usage{
prepare_vars(country, data)
}
\arguments{
\item{country}{a vector of ISO code(s) of the focal country or countries}

\item{data}{the complete dataset}
}
\description{
Identify variables available for focal countries
}
\examples{
prepare_vars(country = "VUT", data = data_rangers)

}
