% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_LMM.R
\name{run_LMM_workflow}
\alias{run_LMM_workflow}
\title{Run the full workflow for LMM}
\usage{
run_LMM_workflow(
  data,
  rerank = TRUE,
  Ncpu = 2,
  coef = 0,
  rep_feature_select = 1000,
  rep_finetune = 1000,
  rep_simu = 10000
)
}
\arguments{
\item{data}{the full dataset}

\item{rerank}{whether or not to recompute variable importance recursively during selection (default = TRUE)}

\item{Ncpu}{the number of CPU cores to be used (default = 1)}

\item{coef}{the coefficient used to population rangers in unsurveyed part of surveyed countries (see \code{\link[=fill_PA_area]{fill_PA_area()}})}

\item{rep_feature_select}{the number of replicates for the feature selection (default = 1000)}

\item{rep_finetune}{the number of replicates for fine tuning (default = 1000)}

\item{rep_simu}{the number of simulation replicates (default = 10000)}
}
\value{
a list with all the output information
}
\description{
Run the full workflow for LMM
}
\examples{
\dontrun{
  LMM_small_test <- run_LMM_workflow(data = data_rangers, Ncpu = 2, coef = 0,
                                     rep_feature_select = 2, rep_finetune = 2, rep_simu = 2)
}

}
