% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_completeness}
\alias{table_completeness}
\title{Create completeness table (numbers)}
\usage{
table_completeness(data, outliers = "GRL")
}
\arguments{
\item{data}{the complete dataset}

\item{outliers}{a vector with the iso code for the countries/territories to discard (default = \code{"GRL"})}
}
\value{
a tibble
}
\description{
This function creates the table that indicate how many countries and territories have ranger staff and other staff documented in our data.
}
\examples{
table_completeness(data_rangers)

}
