% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_completeness_vars}
\alias{table_completeness_vars}
\title{Create completeness table (variables)}
\usage{
table_completeness_vars(data)
}
\arguments{
\item{data}{the complete dataset}
}
\value{
a tibble
}
\description{
This function creates the table that indicate for each variable, how much information is available.
}
\examples{
table_completeness_vars(data_rangers)

}
