% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_core_data}
\alias{table_core_data}
\title{Create core data table}
\usage{
table_core_data(data)
}
\arguments{
\item{data}{the complete dataset created with \href{fetch_data}{\code{fetch_data}()}.}
}
\value{
a tibble
}
\description{
This function creates the table with the core information about surveyed countries/territories.
}
\examples{
table_core_data(data_rangers)

}
